Move 'package to module' intention into intentions package
